Provide non-standard method to centre strings
As indirect * width is not available, and is commonly used for centring text. An alternate method is provided.
If [width] is specified and [.precision] is .0 then %s types will be centred within width. Favouring the left if unable to centre.

Caution. If the application generates non-literal format strings at runtime, then %[width].0s will NOT print 0 characters.