better handling of hex and bin
the print_dec print_hex and print_bin functions have been replaced with print_int. The width and precision behaves the same for all, and is compatible with printf(). Before this change, hex or bin values would default to a width determined by the type, ie. PRIu16 would be 4 digits. This made it difficult to output something like 3FE, which a human probably doesn't want to see anyway, but a machine might.
 The only problem with this change, is that a larger txt buffer is required to store the binary value (64 digits).